dashboards/streamlit-analytics-dashboard-v2/requirements.txt
streamlit>=1.25
pandas>=2.1
supabase>=1.0.0
altair>=5.2
reportlab>=4.0
python-dotenv>=1.0
psycopg2-binary>=2.9
altair_saver>=0.6.0
selenium>=4.0.0
